2008-10-13  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Fixed compile warning on linux-amd64
	* src/FtpConnection.cc
	* src/IteratableChunkChecksumValidator.cc
	* src/MultiDiskAdaptor.cc

	Fixed the bug that unit test fails on linux-amd64
	* test/UtilTest.cc

+++ src/IteratableChunkChecksumValidator.cc
@@ -155,7 +155,7 @@ std::string IteratableChunkChecksumValidator::digest(off_t offset, size_t length
 		      strerror(errno)).str());
     }
     size_t wlength;
-    if(max < curoffset+r) {
+    if(max < static_cast<off_t>(curoffset+r)) {
       wlength = max-curoffset-woffset;
     } else {
       wlength = r-woffset;
